# backend/src/main/java/grimsi/accservermanager/backend/dto/ErrorDto.java
package grimsi.accservermanager.backend.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
public class ErrorDto {
    private int code;
    private String message;
}