# backend/src/main/java/grimsi/accservermanager/backend/dto/UserDto.java
package grimsi.accservermanager.backend.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import javax.validation.constraints.NotBlank;
import javax.validation.constraints.NotNull;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class UserDto {

    @NotBlank(message = "username is required.")
    private String username;

    @NotNull(message = "password is required.")
    private String password;
}